Relocating datetime_mean to time_util + fixing tests.
